Use a service to sync with server
