+++ app/src/main/AndroidManifest.xml
@@ -5,8 +5,8 @@
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
-    <!-- Ensure this permission is only used in older Android versions,
-    After API 18 you don't need this permission, you get an app cache by default
+    <!-- This permission is only used in older Android versions,
+    After API 18 you don't need it, you get an app cache by default
     -->
     <uses-permission
         android:name="android.permission.WRITE_EXTERNAL_STORAGE"
@@ -54,11 +54,14 @@
             </intent-filter>
         </service>
 
-        <service android:name=".service.UpdateInstanceIDListenerService">
+        <service android:name=".service.TokenRefreshedService">
             <intent-filter>
                 <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
             </intent-filter>
         </service>
+        <service
+            android:name=".service.SyncTokenService"
+            android:exported="false" />
 
         <activity
             android:name=".AddAccountActivity"
