+++ app/src/main/java/fr/neamar/lolgamedata/LolApplication.java
@@ -10,11 +10,6 @@ import android.util.Log;
 
 import com.amplitude.api.Amplitude;
 import com.amplitude.api.AmplitudeClient;
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.google.firebase.iid.FirebaseInstanceId;
 import com.mixpanel.android.mpmetrics.MixpanelAPI;
 import com.nostra13.universalimageloader.cache.memory.impl.WeakMemoryCache;
 import com.nostra13.universalimageloader.core.DisplayImageOptions;
@@ -22,16 +17,10 @@ import com.nostra13.universalimageloader.core.ImageLoader;
 import com.nostra13.universalimageloader.core.ImageLoaderConfiguration;
 
 import org.json.JSONArray;
-import org.json.JSONObject;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
-import java.util.ArrayList;
 import java.util.List;
 
-import fr.neamar.lolgamedata.network.VolleyQueue;
 import fr.neamar.lolgamedata.pojo.Account;
-import fr.neamar.lolgamedata.volley.NoCacheRetryJsonRequest;
 
 public class LolApplication extends Application {
     private static final String TAG = "LolApplication";
@@ -123,60 +112,5 @@ public class LolApplication extends Application {
 
     public String getApiUrl() {
         return getString(R.string.API_URL);
-    }
-
-    /*
-        Synchronizes Firebase token with the main user account
-     */
-    public void syncTokenToServer() {
-        AccountManager accountManager = new AccountManager(this);
-        ArrayList<Account> accounts = accountManager.getAccounts();
-        if (accounts.isEmpty()) {
-            Log.i(TAG, "No account yet, skipping token registration");
-            return;
-        }
-        final Account account = accounts.get(0);
-
-        final String token = FirebaseInstanceId.getInstance().getToken();
-
-        if(token == null || token.isEmpty()) {
-            Log.i(TAG, "Firebase token not ready yet, skipping token registration");
-            return;
-        }
-
-        // We have both a token and an account, send that to the server
-        // Instantiate the RequestQueue.
-        final RequestQueue queue = VolleyQueue.newRequestQueue(this);
-        String url;
-        try {
-            url = getApiUrl() + "/push?token=" + token + "&summoner=" + URLEncoder.encode(account.summonerName, "UTF-8") + "&region=" + account.region;
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-            return;
-        }
-
-        NoCacheRetryJsonRequest jsonRequest = new NoCacheRetryJsonRequest(Request.Method.GET, url,
-                null,
-                new Response.Listener<JSONObject>() {
-                    @Override
-                    public void onResponse(JSONObject response) {
-                        Log.i(TAG, "Token registered with server for user " + account.summonerName);
-                        queue.stop();
-                    }
-                }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                Log.e(TAG, error.toString());
-
-                try {
-                    String responseBody = new String(error.networkResponse.data, "utf-8");
-                    Log.i(TAG, responseBody);
-                } catch (UnsupportedEncodingException | NullPointerException e) {
-                    e.printStackTrace();
-                }
-            }
-        });
-
-        queue.add(jsonRequest);
     }
 }
